refs #8733. Check existance before deletion.
When we have group workspaces, if the parent group workspace is deleted first, then it will also remove the children. If that happens, the names previously fetched (which also include the childeren) may be invalid. Checking each workspace before deletion will avoid the invalid delete.

+++ Code/Mantid/scripts/Reflectometry/isis_reflectometry/quick.py
@@ -344,13 +344,13 @@ def polCorr(polcorr, IvsLam, crho, calpha, cAp, cPp):
         print message
     return IvsLam
     
+
 def cleanup():
     names = mtd.getObjectNames()
     for name in names:
-        if re.search("^_", name):
+        if re.search("^_", name) and mtd.doesExist(name):
             print "deleting " + name
             DeleteWorkspace(name)
-    
     
 def get_defaults(run_ws, polcorr = False):
     '''
